--- 64 bit *: adding x64 platform to pom *: Allocator.Allocate should use size_t as type (64 bit)

+++ source/main/include/xunittest/private/ut_Config.h
@@ -8,7 +8,7 @@ namespace UnitTest
 	public:
 		virtual				~Allocator() {}
 
-		virtual void*		Allocate(int size) = 0;
+		virtual void*		Allocate(size_t size) = 0;
 		virtual void		Deallocate(void* ptr) = 0;
 	};
 
@@ -37,9 +37,9 @@ namespace UnitTest
 	public:
 							NullAllocator() { }
 
-		virtual void*		Allocate(int size) { return 0; }
+		virtual void*		Allocate(size_t size) { return 0; }
 		virtual void		Deallocate(void* ptr) {}
-		void	Release()		{}
+		void				Release()		{}
 	};
